(NPUP-15) Port left shift operator to the dispatcher.
Porting the left shift operator's implementation to the dispatcher.
